+++ barc_blanket/utilities.py
@@ -1,8 +1,8 @@
 import os
 from contextlib import contextmanager
 
-CROSS_SECTIONS = '/usr/local/share/xs_data/endfb80_hdf5/cross_sections.xml'
-CHAIN_FILE = '/home/zkeith/openmc_resources/chain_endfb80_sfr.xml'
+CROSS_SECTIONS = '/home/hallj/endfb71_hdf5/cross_sections.xml'
+CHAIN_FILE = '/usr/local/share/xs_data/depletion_chains/chain_endfb80_pwr.xml'
 
 @contextmanager
 def working_directory(directory):
